requirements.txt
joblib==1.4.2
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.3
